requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
antlr4-python3-runtime==4.9.3
anyio==4.13.0
certifi==2026.2.25
click==8.3.1
cloudpickle==3.1.2
cuda-bindings==13.2.0
cuda-pathfinder==1.5.0
cuda-toolkit==13.0.2
einops==0.8.2
fairscale==0.4.13
filelock==3.25.2
fsspec==2026.2.0
ftfy==6.1.1
fvcore==0.1.5.post20221221
grpcio==1.78.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
hydra-core==1.3.2
idna==3.11
ImageIO==2.37.3
iopath==0.1.10
Jinja2==3.1.6
joblib==1.5.3
lazy-loader==0.5
lightning-utilities==0.15.3
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==1.26.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
omegaconf==2.3.0
packaging==26.0
pillow==12.1.1
portalocker==3.2.0
protobuf==7.34.1
psutil==7.2.2
pycocotools==2.0.11
Pygments==2.19.2
PyYAML==6.0.3
regex==2026.2.28
rich==14.3.3
safetensors==0.7.0
-e git+https://kapilparab@github.com/kapilparab/sam3-UA.git@b3a4cf404683b3c9b293cb10d1a2923bb17ffa6a#egg=sam3
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
shellingham==1.5.4
submitit==1.5.4
sympy==1.14.0
tabulate==0.10.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.3.0
threadpoolctl==3.6.0
tifffile==2026.3.3
timm==1.0.26
torch==2.11.0
torchmetrics==1.9.0
torchvision==0.26.0
tqdm==4.67.3
triton==3.6.0
typer==0.24.1
typing_extensions==4.15.0
wcwidth==0.6.0
Werkzeug==3.1.7
yacs==0.1.8
zstandard==0.25.0
